+++ libs/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.37.1
+
+### Patch Changes
+
+- 8c47629: **Dropdown:** Fix regression in transitional styles
+
+## 1.37.0
+
+### Minor Changes
+
+- 36ade3b: **New component:** Add Floating Action Button
+
 ## 1.36.0
 
 ### Minor Changes